Change suspected culprut of Firefox dark theme bug
There seems to be a bug on Firefox where the dark theme kinda flashes
when navigating. Though I'm not able to reproduce on local. I suspect
it's the favicon which for some reason I decided to animate. I guess I
thought a beating heart would be cute. Which it is. But not at what
cost?!